Add files via upload
Welcome to my COVID-19 SQL Data Analytics Portfolio project, an in-depth compilation of SQL queries, scripts, and database designs aimed at exploring and analyzing the multifaceted impact of the COVID-19 pandemic. Through this project, I showcase my expertise in utilizing SQL to extract meaningful insights from large datasets and contribute to a better understanding of the pandemic's effects.

Project Overview:
The COVID-19 pandemic has generated an immense amount of data spanning infection rates, testing, hospitalizations, vaccinations, and more. This portfolio project is a testament to my commitment to using SQL as a powerful tool to harness this data and derive actionable insights.